post-share-delete file with s3 storage

+++ LinkVerse/post-service/src/main/java/com/LinkVerse/post/service/S3Service.java
@@ -1,17 +1,14 @@
 package com.LinkVerse.post.service;
 
 
+import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.amazonaws.services.s3.model.S3Object;
-import com.amazonaws.services.s3.model.S3ObjectInputStream;
-import com.amazonaws.util.IOUtils;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -45,24 +42,33 @@ public class S3Service {
         return fileUrls;
     }
 
-    public byte[] downloadFile(String fileName) {
-        S3Object s3Object = s3Client.getObject(bucketName, fileName);
-        S3ObjectInputStream inputStream = s3Object.getObjectContent();
-        try {
-            byte[] content = IOUtils.toByteArray(inputStream);
-            return content;
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-
     public String deleteFile(String fileName) {
-        s3Client.deleteObject(bucketName, fileName);
-        return fileName + " removed ...";
+        try {
+            boolean exists = s3Client.doesObjectExist(bucketName, fileName);
+            if (exists) {
+                s3Client.deleteObject(bucketName, fileName);
+                return fileName + " removed from S3 successfully.";
+            } else {
+                log.error("File does not exist: {}", fileName);
+                return "File does not exist on S3.";
+            }
+        } catch (SdkClientException e) {
+            log.error("Error occurred while deleting file from S3: {}", e.getMessage());
+            return "Error occurred while deleting file from S3.";
+        }
     }
 
+//    public byte[] downloadFile(String fileName) {
+//        S3Object s3Object = s3Client.getObject(bucketName, fileName);
+//        S3ObjectInputStream inputStream = s3Object.getObjectContent();
+//        try {
+//            byte[] content = IOUtils.toByteArray(inputStream);
+//            return content;
+//        } catch (IOException e) {
+//            e.printStackTrace();
+//        }
+//        return null;
+//    }
 
     private File convertMultiPartFileToFile(MultipartFile file) {
         File convertedFile = new File(file.getOriginalFilename());
